A starting base with an empty path class.
Two starting controllers. One is simply responsible for the gyroscope
feedback the other is a pathfollower which runs a control thread and
pushes the motion profile.

Last class is a motion profile template which the pathfollower takes.

In team67.robot there is a simple robot.java, some constants including
configuration of PIDFs, and a Paths class and package to put the actual
profiles in.